# api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Shopmon API
  description: Shopware monitoring application API
  version: 1.0.0
servers:
  - url: /api
    description: API base path
...
security:
  - bearerAuth: []
paths:
...
  /organizations/{orgId}/projects:
    get:
      operationId: getOrganizationProjects
      tags: [Projects]
      summary: Get all projects in an organization
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createProject
      tags: [Projects]
      summary: Create a new project
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /organizations/{orgId}/projects/{projectId}:
    patch:
      operationId: updateProject
      tags: [Projects]
      summary: Update a project
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      operationId: deleteProject
      tags: [Projects]
      summary: Delete a project
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Project deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────
  # API Keys
  # ──────────────────────────────────────────────
...
  /packages-token/configuration:
    get:
      operationId: getPackagesTokenConfiguration
      tags: [PackagesToken]
      summary: Get packages token configuration
      responses:
        "200":
          description: Packages token configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackagesTokenConfiguration"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /organizations/{orgId}/projects/{projectId}/packages-tokens:
    get:
      operationId: getPackagesTokens
      tags: [PackagesToken]
      summary: List packages tokens for a project
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: List of packages tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackagesToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createPackagesToken
      tags: [PackagesToken]
      summary: Create a new packages token
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackagesTokenRequest"
      responses:
        "201":
          description: Packages token created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /organizations/{orgId}/projects/{projectId}/packages-tokens/{tokenId}:
    delete:
      operationId: deletePackagesToken
      tags: [PackagesToken]
      summary: Delete a packages token
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/TokenId"
      responses:
        "204":
          description: Packages token deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{orgId}/projects/{projectId}/packages-tokens/{tokenId}/sync:
    post:
      operationId: syncPackagesToken
      tags: [PackagesToken]
      summary: Sync a packages token
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/TokenId"
      responses:
        "204":
          description: Sync triggered
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────
  # Deployments
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication

  # ──────────────────────────────────────────────
  # Parameters
  # ──────────────────────────────────────────────
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
      description: Organization ID
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
      description: Project ID
...
    TokenId:
      name: tokenId
      in: path
      required: true
      schema:
        type: integer
      description: Packages token ID
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ──────────────────────────────────────────────
  # Schemas
  # ──────────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    # ── Account ──────────────────────────────────
...
    Project:
      type: object
      required:
        - id
        - name
        - description
        - gitUrl
        - organizationId
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        gitUrl:
          type: string
          nullable: true
        organizationId:
          type: string
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        gitUrl:
          type: string
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        gitUrl:
          type: string

    # ── API Keys ─────────────────────────────────
...
    PackagesTokenConfiguration:
      type: object
      required:
        - configured
        - composerUrl
      properties:
        configured:
          type: boolean
        composerUrl:
          type: string
          nullable: true
    PackagesToken:
      type: object
      required:
        - id
        - source
        - lastSyncedAt
      properties:
        id:
          type: integer
        source:
          type: string
        lastSyncedAt:
          type: string
          nullable: true
          format: date-time
    CreatePackagesTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string

    # ── Deployments ──────────────────────────────
...